requirements.txt
# --- Web Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.0
python-dotenv>=1.0.1
httpx>=0.27.0  # [추가] API 요청용

# --- Whisper & Diarization ---
faster-whisper==1.0.3
pyannote.audio==3.1.1

# --- Common Utils ---
numpy>=1.24
soundfile==0.12.1

# --- PyTorch (CPU build for simple local setup) ---
torch==2.4.1
torchaudio==2.4.1
# pip3 install torch torchaudio --index-url https://download.pytorch.org/whl/cu124

# --- AI & Agent (MCP 핵심) [추가됨] ---
langchain-ollama>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
langchain>=0.3.0

# --- Common Utils ---
numpy>=1.24
soundfile>=0.12.1